fix bugs

+++ src/vagons/vagons.service.ts
@@ -18,13 +18,24 @@ export class VagonsService {
 
 
 
-  async create(dto: CreateVagonDto): Promise<Vagon> {
+  async create(dto: CreateVagonDto, vchdId?: string, role?: string): Promise<Vagon> {
     let vchd: Vchd;
+    if(role === 'superadmin'){ ///keyinro to'g'irlanadi
+      throw new BadRequestException('Superadmin cannot create vagons directly. Please use admin role.');
+    }
+    if (!vchdId) {
+      throw new BadRequestException('VCHD ID is required');
+    }
 
     try {
-      vchd = await this.vchdRepo.findOneByOrFail({ id: dto.vchdId });
+      vchd = await this.vchdRepo.findOneByOrFail({ id: vchdId });
     } catch (error) {
-      throw new NotFoundException(`VCHD not found with id: ${dto.vchdId}`);
+      throw new NotFoundException(`VCHD not found with id: ${vchdId}`);
+    }
+    
+    const existingVagon = await this.vagonRepo.findOneBy({ number: dto.number });
+    if (existingVagon) {
+      throw new BadRequestException(`Vagon with number ${dto.number} already exists`);
     }
 
     let importedTime: Date;
@@ -62,8 +73,11 @@ export class VagonsService {
     return this.vagonRepo.save(vagon);
   }
 
-  async findAll(): Promise<Vagon[]> {
-    return this.vagonRepo.find();
+  async findAll(vchdId: string): Promise<Vagon[]> {
+    return this.vagonRepo.find({
+      where: { vchd: { id: vchdId } },
+      order: { importedTime: 'DESC' },
+    });
   }
 
   async findOne(id: string): Promise<Vagon> {
@@ -72,10 +86,13 @@ export class VagonsService {
     return vagon;
   }
 
-  async update(id: string, dto: UpdateVagonDto): Promise<Vagon> {
+  async update(id: string, dto: UpdateVagonDto, vchdId: string, role: string): Promise<Vagon> {
+    if(role === 'superadmin') {
+      throw new BadRequestException('Superadmin cannot update vagons directly. Please use admin role.');
+    }
     const vagon = await this.findOne(id);
-    if (dto.vchdId) {
-      const vchd = await this.vchdRepo.findOneBy({ id: dto.vchdId });
+    if (vchdId) {
+      const vchd = await this.vchdRepo.findOneBy({ id: vchdId });
       if (!vchd) throw new NotFoundException('VCHD not found');
       vagon.vchd = vchd;
     }
@@ -84,8 +101,15 @@ export class VagonsService {
     return this.vagonRepo.save(vagon);
   }
 
-  async remove(id: string): Promise<void> {
+  async remove(id: string, vchdId: string, role:string): Promise<void> {
+    if(role === 'superadmin') {
+      throw new BadRequestException('Superadmin cannot delete vagons directly. Please use admin role.');
+    }
     const vagon = await this.findOne(id);
+    if (vagon.vchd.id !== vchdId) {
+      throw new BadRequestException('Vagon does not belong to the specified VCHD');
+    }
+    if (!vagon) throw new NotFoundException('Vagon not found');
     await this.vagonRepo.remove(vagon);
   }
